refactor(CarGenerator): CarFactory -> CarGenerator 이동
로직상 domain -> controller
