+++ src/test/java/javaracingcar/controller/CarGeneratorTest.java
@@ -1,6 +1,12 @@
-package javaracingcar.domain;
+package javaracingcar.controller;
 
+import javaracingcar.controller.CarGenerator;
+import javaracingcar.domain.Car;
 import org.junit.jupiter.api.Test;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.*;
@@ -8,30 +14,30 @@ import static org.junit.jupiter.api.Assertions.*;
 class CarFactoryTest {
     @Test
     void generateCarWithName_null() {
-        assertThatThrownBy(() -> CarFactory.generateCar(null))
+        assertThatThrownBy(() -> CarGenerator.generateCar(null))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessageContaining("null");
     }
 
     @Test
     void generateCarWithName_공백() {
-        assertThatThrownBy(() -> CarFactory.generateCar(""))
+        assertThatThrownBy(() -> CarGenerator.generateCar(""))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessageContaining("빈 문자열");
-        assertThatThrownBy(() -> CarFactory.generateCar("   "))
+        assertThatThrownBy(() -> CarGenerator.generateCar("   "))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessageContaining("빈 문자열");
     }
 
     @Test
     void generateCarWithName_최대길이이하() {
-        Car car = CarFactory.generateCar("Benz");
+        Car car = CarGenerator.generateCar("Benz");
         assertEquals(car.getName(), "Benz");
     }
 
     @Test
     void generateCarWithName_최대길이초과() {
-        assertThatThrownBy(() -> CarFactory.generateCar("Tuscan"))
+        assertThatThrownBy(() -> CarGenerator.generateCar("Tuscan"))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessageContaining("최대길이를 초과");
     }
